dashboard-cloning-tool/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pyjwt==2.8.0
bcrypt==4.1.3
python-multipart==0.0.9
redshift-connector==2.1.1
tableauserverclient==0.28
aiofiles==23.2.1
jinja2==3.1.4
httpx==0.27.0
